--- a/scene/register_scene_types.cpp
+++ b/scene/register_scene_types.cpp
@@ -538,6 +538,7 @@ void register_scene_types() {
 	ObjectTypeDB::register_virtual_type<Material>();
 	ObjectTypeDB::register_type<FixedMaterial>();
 	ObjectTypeDB::register_type<ShaderMaterial>();
+	ObjectTypeDB::register_type<MultiPassMaterial>();
 	ObjectTypeDB::register_type<RoomBounds>();
 	ObjectTypeDB::register_type<MaterialShaderGraph>();
 	ObjectTypeDB::register_type<MaterialShader>();