… bin/tests/test_misc.cpp
… doc/base/classes.xml
… drivers/gles2/rasterizer_gles2.cpp
… scene/resources/material.cpp
… servers/visual/rasterizer_dummy.cpp
… servers/visual/rasterizer_dummy.h
… servers/visual/visual_server_raster.cpp
… servers/visual/visual_server_raster.h
… servers/visual/visual_server_wrap_mt.h
… servers/visual_server.h
… tools/editor/icons/icon_multi_pass_material.png
… tools/editor/plugins/script_editor_plugin.cpp
… tools/editor/plugins/spatial_editor_plugin.cpp