Okay, this PR retools how materials are handled internally somewhat. Materials now include the concept of having multiple 'passes', which contain all the information originally containined in the root material. This is a concept I've seen in many renderers including Ogre3D and Unity's ShaderLab system (which also both include the concept of a fallback mechanism for unsupported shader features; may be worth considering something similar in the future). What this means is that the renderer can render the same instance multiple times but with different shaders and render parameters. The easiest to describe use-case for something like this is rendering an object with an outline using an inverse hull, without having to do the rather cheap hack of just rendering multiple versions of the same geometry on the editor side of things.

Internally on the renderer side of things, not much has changed besides the fact that things are sorted via passes rather than materials now. More could probably be done to provide greater optimization advantages to a multiple pass material system (for example, it may be worth adding a mechanism of skeletal and morph modifications to be cached between multiple passes using the same material on the same instance), but it shouldn't incur any significant penalities to pre-existing functionality, and should also retain backwards compatibility with pre-existing content.

On the editor side of things, the most notable inclusion is that of the new multi-pass material type which allows you to set the number of passes, and then shaders and params for each pass. FixedMaterials and ShaderMaterials still remain and should work exactly as they did before. The only thing I was not able to keep as backwards compatible was changes to the VisualServer frontend due to needing to add new arguments for creating materials, but I don't imagine many people interacting with the VisualServer directly and it's very trivial to change such code. I also just pushed a PR for a minor fix for the tree gui control, and this is needed due the highly dynamic nature of the multipass material's parameter. On the whole, there are probably a lot more interesting things which can be done with this feature beyond the scope of what I had in mind, but this at least should provide a competent self-explanatory frontend to remove the need for a lot of rather inelegant hacks for certain graphical features.

Again, if there are any questions about this PR, please ask me.