@reduz 2016-03-17
I'll check as soon as I can!
On Mar 17, 2016 16:33, "SaracenOne" <notifications@github.com> wrote:

> Okay, this PR retools how materials are handled internally somewhat.
> Materials now include the concept of having multiple 'passes', which
> contain all the information originally containined in the root material.
> This is a concept I've seen in many renderers including Ogre3D and Unity's
> ShaderLab system (which also both include the concept of a fallback
> mechanism for unsupported shader features; may be worth considering
> something similar in the future). What this means is that the renderer can
> render the same instance multiple times but with different shaders and
> render parameters. The easiest to describe use-case for something like this
> is rendering an object with an outline using an inverse hull, without
> having to do the rather cheap hack of just rendering multiple versions of
> the same geometry on the editor side of things.
>
> Internally on the renderer side of things, not much has changed besides
> the fact that things are sorted via passes rather than materials now. More
> could probably be done to provide greater optimization advantages to a
> multiple pass material system (for example, it may be worth adding a
> mechanism of skeletal and morph modifications to be cached between multiple
> passes using the same material on the same instance), but it shouldn't
> incur any significant penalities to pre-existing functionality, and should
> also retain backwards compatibility with pre-existing content.
>
> On the editor side of things, the most notable inclusion is that of the
> new multi-pass material type which allows you to set the number of passes,
> and then shaders and params for each pass. FixedMaterials and
> ShaderMaterials still remain and should work exactly as they did before.
> The only thing I was not able to keep as backwards compatible was changes
> to the VisualServer frontend due to needing to add new arguments for
> creating materials, but I don't imagine many people interacting with the
> VisualServer directly and it's very trivial to change such code. I also
> just pushed a PR for a minor fix for the tree gui control, and this is
> needed due the highly dynamic nature of the multipass material's parameter.
> On the whole, there are probably a lot more interesting things which can be
> done with this feature beyond the scope of what I had in mind, but this at
> least should provide a competent self-explanatory frontend to remove the
> need for a lot of rather inelegant hacks for certain g raphical features.
>
> Again, if there are any questions about this PR, please ask me.
> ------------------------------
> You can view, comment on, or merge this pull request online at:
>
>   https://github.com/godotengine/godot/pull/4084
> Commit Summary
>
>    - Multipass material support.
>
> File Changes
>
>    - *M* bin/tests/test_misc.cpp
>    <https://github.com/godotengine/godot/pull/4084/files#diff-0> (6)
>    - *M* drivers/gles2/rasterizer_gles2.cpp
>    <https://github.com/godotengine/godot/pull/4084/files#diff-1> (563)
>    - *M* drivers/gles2/rasterizer_gles2.h
>    <https://github.com/godotengine/godot/pull/4084/files#diff-2> (149)
>    - *M* modules/gridmap/grid_map_editor_plugin.cpp
>    <https://github.com/godotengine/godot/pull/4084/files#diff-3> (24)
>    - *M* scene/register_scene_types.cpp
>    <https://github.com/godotengine/godot/pull/4084/files#diff-4> (1)
>    - *M* scene/resources/material.cpp
>    <https://github.com/godotengine/godot/pull/4084/files#diff-5> (510)
>    - *M* scene/resources/material.h
>    <https://github.com/godotengine/godot/pull/4084/files#diff-6> (124)
>    - *M* servers/visual/rasterizer.cpp
>    <https://github.com/godotengine/godot/pull/4084/files#diff-7> (30)
>    - *M* servers/visual/rasterizer.h
>    <https://github.com/godotengine/godot/pull/4084/files#diff-8> (28)
>    - *M* servers/visual/rasterizer_dummy.cpp
>    <https://github.com/godotengine/godot/pull/4084/files#diff-9> (120)
>    - *M* servers/visual/rasterizer_dummy.h
>    <https://github.com/godotengine/godot/pull/4084/files#diff-10> (89)
>    - *M* servers/visual/visual_server_raster.cpp
>    <https://github.com/godotengine/godot/pull/4084/files#diff-11> (61)
>    - *M* servers/visual/visual_server_raster.h
>    <https://github.com/godotengine/godot/pull/4084/files#diff-12> (29)
>    - *M* servers/visual/visual_server_wrap_mt.h
>    <https://github.com/godotengine/godot/pull/4084/files#diff-13> (29)
>    - *M* servers/visual_server.cpp
>    <https://github.com/godotengine/godot/pull/4084/files#diff-14> (6)
>    - *M* servers/visual_server.h
>    <https://github.com/godotengine/godot/pull/4084/files#diff-15> (28)
>    - *M* tools/editor/editor_plugin_settings.cpp
>    <https://github.com/godotengine/godot/pull/4084/files#diff-16> (2)
>    - *A* tools/editor/icons/icon_multi_pass_material.png
>    <https://github.com/godotengine/godot/pull/4084/files#diff-17> (0)
>    - *M* tools/editor/plugins/script_editor_plugin.cpp
>    <https://github.com/godotengine/godot/pull/4084/files#diff-18> (2)
>    - *M* tools/editor/plugins/spatial_editor_plugin.cpp
>    <https://github.com/godotengine/godot/pull/4084/files#diff-19> (6)
>
> Patch Links:
>
>    - https://github.com/godotengine/godot/pull/4084.patch
>    - https://github.com/godotengine/godot/pull/4084.diff
>
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly or view it on GitHub
> <https://github.com/godotengine/godot/pull/4084>
>