@SaracenOne 2016-07-18
Speaking of which, I'm just about to push some documentation updates to this PR.