@SuperUserNameMan 2016-09-05
I'd be interested too see this PR merged to 2.2 too.