@SuperUserNameMan 2016-09-06
I managed to test this MultiPassMaterial with fur-shading rendering.

First thing I want to say is that Multi pass material gives a great boost of performances for fur-shading. **The rendering went from 30 fps to 90 fps.**
Also, with this MultiPassMaterial, you don't have to regenerate the geometry of the Mesh for each layer of fur.
Instead, you simply add as much passes as layers to your material with a different fur_depth paramter for each, and you tell you Vertex shader to regenerate the geometry for each pass in realtime.
This is a gain of 5 to 10 seconds !

Now, the second thing is about the problem I've found with this MultiPassMaterial implementation :

1) it seems that each passes is trying to connect to the same signal 'changed' : 
````
ERROR: Signal 'changed'' already connected to given method '_shader_changed' in
that object.
   At: core/object.cpp:1436
````
2) it seems that the order in which each pass is rendered is not 100% reliable. Maybe related to z-sorting with transparent material.

If you want the multi-pass version of my fur-demo to test more, just tell me.