@SuperUserNameMan 2016-09-07
for those willing to test without having to struggle with git and compilers, i've put a Godot 2.1 win32 that includes this PR into here :
https://github.com/SuperUserNameMan/GodotFurShaderDemo/tree/master/with_MultiPassMaterial

You can also find the godot project for multi pass fur-shading at the same palce