@LeonardMeagher2 2016-09-20
Will this be a thing someone can do in 3.0 with the new shading language?